+++ common.hpp
@@ -141,6 +141,11 @@ constexpr const char* error_caption ("*** ERROR ***");
   { errorExit (msg. c_str ()); }
 
 
+
+class Notype {};
+
+
+
 struct Nocopy
 {
 protected:
@@ -493,6 +498,9 @@ inline bool isDelimiter (char c)
            && ! isLetter (c);
   }
   
+inline bool isSpace (char c)
+  { return c > '\0' && c <= ' ' && isspace (c); }
+
 inline string uchar2hex (uchar c)
   { constexpr char hex [16] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
     string res ("  ");
@@ -602,6 +610,11 @@ public:
 
 
 // STL algorithms
+
+template <typename T>
+  inline bool lessPtr (const T* x,
+                       const T* y)
+    { return *x < *y; }
 
 template <typename T>
   inline void sort (T &t)
@@ -956,9 +969,6 @@ inline string prepend (const string &prefix,
   	  return noString;
   	return prefix + s;
   }
-
-inline bool isSpace (char c)
-  { return c > '\0' && c <= ' ' && isspace (c); }
 
 bool strBlank (const string &s);
 
@@ -1373,12 +1383,6 @@ template <typename T>
 
 //
 
-bool verbose (int inc = 0);
-
-int getVerbosity ();
-
-
-
 void exec (const string &cmd,
            const string &logFName = string());
 
@@ -1476,7 +1480,13 @@ template <typename Func, typename Res, typename... Args>
 
 
 
-//
+// Verbosity
+
+bool verbose (int inc = 0);
+
+int getVerbosity ();
+
+
 
 class Verbose
 {
@@ -1494,6 +1504,8 @@ public:
     { return isMainThread () /*Threads::empty ()*/; }
 };
 
+
+
 struct Unverbose 
 {
 	Unverbose ();
@@ -1502,13 +1514,11 @@ struct Unverbose
   
 
 
+
+//
+
 struct Json;
 struct JsonContainer;
-
-
-
-class Notype {};
-
 
 
 
@@ -2194,6 +2204,8 @@ public:
     { delete (*this) [index];
       P::eraseAt (index);
     }
+  void sortPtr ()
+    { P::sort (lessPtr<T>); }
   void sortBubblePtr ()
     { FOR_START (size_t, i, 1, P::size ())
 		    FOR_REV (size_t, j, i)
@@ -3366,158 +3378,7 @@ public:
 
 
 
-
-struct TextTable : Named
-// Tab-delimited (tsv) table with a header
-// name: file name or empty()
-{
-  bool pound {false};
-    // '#' in the beginning of header
-  bool saveHeader {true};
-  struct Header : Named
-  { 
-    size_t len_max {0};
-    // Type
-    bool numeric {true};
-    // Valid if numeric
-    bool scientific {false};
-    streamsize decimals {0};
-    bool null {false};
-    explicit Header (const string &name_arg)
-      : Named (name_arg)
-      {}
-    void qc () const override;
-    void saveText (ostream& os) const override
-      { os         << name 
-           << '\t' << len_max 
-           << '\t' << (numeric ? ((scientific ? "float" : "int") + string ("(") + to_string (decimals) + ")") : "char") 
-           << '\t' << (null ? "null" : "not null"); 
-      }
-  };
-  Vector<Header> header;
-    // size() = number of columns
-  Vector<StringVector> rows;
-    // StringVector::size() = header.size()
-  typedef  size_t  ColNum;
-    // no_index <=> no column
-  typedef  size_t  RowNum;
-    // no_index <=> no row
-    
-    
-  struct Error : runtime_error
-  {
-    Error (const TextTable &tab,
-           const string &what)
-      : runtime_error (what + "\nIn table file: " + tab. name)
-      {}
-  };
-    
-
-  explicit TextTable (const string &tableFName,
-                      const string &columnSynonymsFName = string());
-    // columnSynonymsFName: syn_format
-  static constexpr const char* syn_format {"Column synonyms file with the format: {<main synonym> <eol> {<synonym> <eol>}* {<eol>|<eof>}}*"};
-  TextTable () = default;
-  TextTable (bool pound_arg,
-             const Vector<Header> &header_arg)
-    : pound (pound_arg)
-    , header (header_arg)
-    {}
-private:
-  void setHeader ();
-public:
-  void qc () const override;
-  void saveText (ostream &os) const override;    
-        
-  
-  static bool getDecimals (string s,
-                           bool &hasPoint,
-                           streamsize &decimals);
-    // Return: true => scientific number
-  void printHeader (ostream &os) const;
-  ColNum col2num_ (const string &columnName) const;
-    // Retuirn: no_index <=> no columnName
-  ColNum col2num (const string &columnName) const
-  { const ColNum i = col2num_ (columnName);
-    if (i == no_index)
-      throw Error (*this, "Cannot find column name " + strQuote (columnName));
-    return i;
-  }
-  Vector<ColNum> columns2nums (const StringVector &columns) const
-    { Vector<ColNum> nums;  nums. reserve (columns. size ());
-      for (const string &s : columns)
-        nums << col2num (s);
-      return nums;
-    }
-  bool hasColumn (const string &columnName) const
-    { return col2num_ (columnName) != no_index; }
-private:
-  int compare (const StringVector& row1,
-               const StringVector& row2,
-               ColNum column) const;
-public:
-  void filterColumns (const StringVector &newColumnNames);
-    // Input: newColumnNames: in header::name's
-    //          can be repeated
-    //          ordered
-  void sort (const StringVector &by);
-  void group (const StringVector &by,
-              const StringVector &sum,
-              const StringVector &aggr);
-    // Invokes: filterColumns(by + sum + aggr)
-private:
-  void merge (RowNum toRowNum,
-              RowNum fromRowNum,
-              const Vector<ColNum> &sum,
-              const Vector<ColNum> &aggr);
-public:
-  void colNumsRow2values (const Vector<ColNum> &colNums,
-                          RowNum row_num,
-                          StringVector &values) const;
-    // Output: values
-  RowNum find (const Vector<ColNum> &colNums,
-               const StringVector &targetValues,
-               RowNum rowNum_start) const;
-  StringVector col2values (ColNum col) const;
-
-
-  struct Key
-  {
-    const Vector<ColNum> colNums;
-    unordered_map<StringVector,RowNum,StringVector::Hasher> data;
-
-    Key (const TextTable &tab,
-         const StringVector &columns);
-         
-    RowNum find (const StringVector &values) const
-      { const auto& it = data. find (values);
-        if (it != data. end ())
-          return it->second;
-        return no_index;
-      }
-  };
-
-
-  struct Index
-  {
-    const Vector<ColNum> colNums;
-    unordered_map<StringVector,Vector<RowNum>,StringVector::Hasher> data;
-
-    Index (const TextTable &tab,
-           const StringVector &columns);
-         
-    const Vector<RowNum>* find (const StringVector &values) const
-      { const auto& it = data. find (values);
-        if (it == data. end ())
-          return nullptr;
-        return & it->second;
-      }
-  };
-};
-
-
 		
-
 // Json
 
 struct JsonNull;
