+++ tsv.cpp
@@ -1,0 +1,621 @@
+// tsv.cpp
+
+/*===========================================================================
+*
+*                            PUBLIC DOMAIN NOTICE                          
+*               National Center for Biotechnology Information
+*                                                                          
+*  This software/database is a "United States Government Work" under the   
+*  terms of the United States Copyright Act.  It was written as part of    
+*  the author's official duties as a United States Government employee and 
+*  thus cannot be copyrighted.  This software/database is freely available 
+*  to the public for use. The National Library of Medicine and the U.S.    
+*  Government have not placed any restriction on its use or reproduction.  
+*                                                                          
+*  Although all reasonable efforts have been taken to ensure the accuracy  
+*  and reliability of the software and data, the NLM and the U.S.          
+*  Government do not and cannot warrant the performance or results that    
+*  may be obtained by using this software or data. The NLM and the U.S.    
+*  Government disclaim all warranties, express or implied, including       
+*  warranties of performance, merchantability or fitness for any particular
+*  purpose.                                                                
+*                                                                          
+*  Please cite the author in any work or product based on this material.   
+*
+* ===========================================================================
+*
+* Author: Vyacheslav Brover
+*
+* File Description:
+*   TSV table
+*
+*/
+
+
+#undef NDEBUG
+#include "common.inc"
+
+#include "tsv.hpp"
+
+
+
+
+namespace Common_sp
+{
+ 
+
+
+// TextTable
+
+void TextTable::Header::qc () const
+{
+  if (! qc_on)
+    return;
+
+  Named::qc ();
+    
+  QC_IMPLY (scientific, numeric);
+  QC_IMPLY (decimals, numeric);
+}
+
+
+
+TextTable::TextTable (const string &tableFName,
+                      const string &columnSynonymsFName)
+: Named (tableFName)
+{  
+  {
+    LineInput f (tableFName);
+    if (! f. nextLine ())
+      throw Error (*this, "Cannot read the table header");
+    if (! f. line. empty () && f. line. front () == '#')
+    {
+      pound = true;
+      f. line. erase (0, 1);
+    }
+    {
+      StringVector h (f. line, '\t', true);
+      for (string& s : h)
+        header << move (Header (move (s)));
+    }
+    while (f. nextLine ())
+    {
+      StringVector line (f. line, '\t', true);
+      rows << move (line);
+      ASSERT (line. empty ());
+    }
+  }
+  
+  if (! columnSynonymsFName. empty ())
+  {
+    LineInput colF (columnSynonymsFName);
+    string mainSyn;
+    while (colF. nextLine ())
+    {
+      trim (colF. line);
+      const string& syn = colF. line;
+      if (syn. empty ())
+        mainSyn. clear ();
+      else
+      {
+        if (mainSyn. empty ())
+          mainSyn = syn;
+        else
+          if (mainSyn != syn)
+          {
+            const ColNum i = col2num_ (syn);
+            if (i != no_index)
+            {
+              if (hasColumn (mainSyn))
+                throw runtime_error ("Table " + strQuote (name) + ": Column " + strQuote (mainSyn) + " already exists");
+              else
+                header [i]. name = mainSyn;
+            }
+          }
+      }
+    }
+  }
+  
+  setHeader ();
+}
+
+
+
+void TextTable::setHeader ()
+{
+  RowNum row_num = 0;
+  for (const StringVector& row : rows)
+  {
+    row_num++;
+    if (row. size () != header. size ())
+      throw Error (*this, "Row " + to_string (row_num) + " contains " + to_string (row. size ()) + " fields whereas header has " + to_string (header. size ()) + " columns");
+    FFOR (RowNum, i, row. size ())
+    {
+      const string& field = row [i];
+      Header& h = header [i];
+      if (field. empty ())
+      {
+        h. null = true;
+        continue;
+      }
+      maximize (h. len_max, field. size ());
+      if (! h. numeric)
+        continue;
+      {
+        char* endptr = nullptr;
+        strtod (field. c_str (), & endptr);
+        if (endptr != field. c_str () + field. size ())
+        {
+          h. numeric = false;
+          h. scientific = false;
+          h. decimals = 0;
+        }
+      }
+      if (h. numeric)
+      {
+        bool hasPoint = false;
+        streamsize decimals = 0;
+        if (getDecimals (field, hasPoint, decimals))
+          h. scientific = true;
+        maximize<streamsize> (h. decimals, decimals);
+      }
+    }
+  }
+
+  // Header::len_max for numeric
+  for (const StringVector& row : rows)
+    FFOR (RowNum, i, row. size ())
+    {
+      const string& field = row [i];
+      if (field. empty ())
+        continue;
+      Header& h = header [i];
+      if (h. numeric)
+      {
+        bool hasPoint = false;
+        streamsize decimals = 0;
+        getDecimals (field, hasPoint, decimals);
+        maximize (h. len_max, field. size () + (size_t) (h. decimals - decimals) + (! hasPoint));
+      }
+    }
+}
+
+
+
+void TextTable::qc () const
+{
+  if (! qc_on)
+    return;
+  if (! name. empty ())
+    Named::qc (); 
+
+  {    
+    StringVector v;  v. reserve (header. size ());
+    FFOR (size_t, i, header. size ())
+    {
+      const Header& h = header [i];
+      try { h. qc (); }
+        catch (const exception &e)
+        {
+          throw runtime_error ("Header column #" + to_string (i + 1) + ": " + e. what ());
+        }
+      v << h. name;
+    }
+    v. sort ();
+    const size_t i = v. findDuplicate ();
+    if (i != no_index)
+      throw Error (*this, "Duplicate column name: " + strQuote (v [i]));
+  }
+  
+  FFOR (RowNum, i, rows. size ())
+  {
+    if (rows [i]. size () != header. size ())
+      throw Error (*this, "Row " + to_string (i + 1) + " contains " + to_string (rows [i]. size ()) + " fields whereas table has " + to_string (header. size ()) + " columns");
+    for (const string& field : rows [i])
+    {
+      if (contains (field, '\t'))
+        throw Error (*this, "Field " + strQuote (header [i]. name) + " of row " + to_string (i + 1) + " contains a tab character");
+      if (contains (field, '\n'))
+        throw Error (*this, "Field " + strQuote (header [i]. name) + " of row " + to_string (i + 1) + " contains an EOL character");
+    }
+  }
+}
+
+
+
+void TextTable::saveText (ostream &os) const
+{ 
+  if (saveHeader)
+  { 
+    if (pound)
+      os << '#';
+    bool first = true;
+    for (const Header& h : header)
+    {
+      if (! first)
+        os << '\t';
+      os << h. name;
+      first = false;
+    }
+    os << endl;
+  }
+  
+  for (const StringVector& row : rows)
+  {
+    save (os, row, '\t');
+    os << endl;
+  }
+}
+
+    
+    
+bool TextTable::getDecimals (string s,
+                             bool &hasPoint,
+                             streamsize &decimals)
+{
+  strUpper (s);
+  const size_t ePos     = s. find ('E');
+  const size_t pointPos = s. find ('.');
+  
+  hasPoint = pointPos != string::npos;
+  
+  decimals = 0;
+  if (ePos == string::npos)
+  {
+    if (hasPoint)
+      decimals = (streamoff) (s. size () - (pointPos + 1));
+  }
+  else
+  {
+    if (hasPoint && ePos > pointPos)
+      decimals = (streamoff) (ePos - (pointPos + 1));
+  }
+  
+  return ePos != string::npos;
+}
+
+    
+  
+void TextTable::printHeader (ostream &os) const
+{
+  FFOR (size_t, i, header. size ())
+  {
+    os << i + 1 << '\t';
+    header [i]. saveText (os);
+    os << endl;
+  }
+}
+
+
+
+TextTable::ColNum TextTable::col2num_ (const string &columnName) const
+{ 
+  FFOR (size_t, i, header. size ())
+    if (header [i]. name == columnName)
+      return i;
+  return no_index;
+}
+
+
+
+void TextTable::duplicateColumn (const string &columnName_from,
+                                 const string &columnName_to)
+{
+  ASSERT (! columnName_to. empty ());
+  const ColNum from = col2num (columnName_from);
+  if (hasColumn (columnName_to))
+    throw runtime_error ("Table already has column " + strQuote (columnName_to));
+  header << header [from];
+  header. back (). name = columnName_to;
+  for (StringVector& row : rows)
+    row << row [from];
+  qc ();
+}
+
+
+
+int TextTable::compare (const StringVector& row1,
+                        const StringVector& row2,
+                        ColNum column) const
+{
+  if (header [column]. numeric)
+  {
+    const double a = stod (row1 [column]);
+    const double b = stod (row2 [column]);
+    if (a < b)
+      return -1;
+    if (a > b)
+      return 1;
+    return 0;
+  }
+  
+  const string& a = row1 [column];
+  const string& b = row2 [column];
+  if (a < b)
+    return -1;
+  if (a > b)
+    return 1;
+
+  return 0;
+}
+
+
+
+void TextTable::filterColumns (const StringVector &newColumnNames)
+{
+  const Vector<ColNum> colNums (columns2nums (newColumnNames));
+
+  {  
+    Vector<Header> newHeader;  newHeader. reserve (colNums. size ());
+    for (const ColNum i : colNums)
+      newHeader << header [i];
+    header = move (newHeader);
+  }
+
+  for (StringVector& row : rows)
+  {
+    StringVector newRow;  newRow. reserve (colNums. size ());
+    for (const ColNum i : colNums)
+      newRow << row [i];
+    row = move (newRow);
+  }
+}
+
+
+
+void TextTable::sort (const StringVector &by)
+{
+  const Vector<ColNum> byIndex (columns2nums (by));
+  
+  const auto lt = [&byIndex,this] (const StringVector &a, const StringVector &b) 
+                    { for (const ColNum i : byIndex) 
+                        switch (this->compare (a, b, i))
+                        { case -1: return true;
+                          case  1: return false;
+                        }
+                      // Tie resolution
+                      FFOR (size_t, i, a. size ())
+                        switch (this->compare (a, b, i))
+                        { case -1: return true;
+                          case  1: return false;
+                        }
+                      return false;
+                    };
+  Common_sp::sort (rows, lt);
+}
+
+
+
+namespace
+{
+  struct ColumnPartitionQC
+  {
+    map<string/*colName*/,string/*operation*/> name2oper;
+    
+    void add (const StringVector &cols,
+              const string &oper)
+      {
+        for (const string& s : cols)
+          if (name2oper [s]. empty ())
+            name2oper [s] = oper;
+          else
+            throw runtime_error ("Column " + strQuote (s) + " is used for operations " + strQuote (name2oper [s]) + " and " + strQuote (oper)); 
+      }
+  };
+}
+
+
+
+void TextTable::group (const StringVector &by,
+                       const StringVector &sum,
+                       const StringVector &minV,
+                       const StringVector &maxV,
+                       const StringVector &aggr)
+{
+  const Vector<ColNum> byIndex   (columns2nums (by));
+  const Vector<ColNum> sumIndex  (columns2nums (sum));
+  const Vector<ColNum> minIndex  (columns2nums (minV));
+  const Vector<ColNum> maxIndex  (columns2nums (maxV));
+  const Vector<ColNum> aggrIndex (columns2nums (aggr));
+  
+  // QC
+  {
+    ColumnPartitionQC cp;
+    cp. add (by, "group by");
+    cp. add (sum, "sum");
+    cp. add (minV, "min");
+    cp. add (maxV, "max");
+    cp. add (aggr, "aggregation");
+  }
+  for (const string& s : sum)
+    if (! header [col2num (s)]. numeric)
+      throw runtime_error ("Summation column " + strQuote (s) + " is not numeric");
+  
+  const auto lt = [&byIndex,this] (const StringVector &a, const StringVector &b) 
+                    { for (const ColNum i : byIndex) 
+                        switch (this->compare (a, b, i))
+                        { case -1: return true;
+                          case  1: return false;
+                        }
+                      // Tie resolution
+                      FFOR (size_t, i, a. size ())
+                        switch (this->compare (a, b, i))
+                        { case -1: return true;
+                          case  1: return false;
+                        }
+                      return false;
+                    };
+  Common_sp::sort (rows, lt);
+  
+  RowNum i = 0;  
+  FFOR_START (RowNum, j, 1, rows. size ())
+  {
+    ASSERT (i < j);
+    if (rows [i]. same (rows [j], byIndex))
+      merge (i, j, sumIndex, minIndex, maxIndex, aggrIndex);
+    else
+    {
+      i++;
+      if (i < j)
+        rows [i] = move (rows [j]);
+    }
+  }
+  if (! rows. empty ())
+    i++;
+
+  ASSERT (rows. size () >= i);
+  FFOR (RowNum, k, rows. size () - i)
+    rows. pop_back ();
+    
+  StringVector newColumns;
+  newColumns << by << sum << minV << maxV << aggr;
+  filterColumns (newColumns);
+}
+
+
+
+void TextTable::merge (RowNum toRowNum,
+                       RowNum fromRowNum,
+                       const Vector<ColNum> &sum,
+                       const Vector<ColNum> &minV,
+                       const Vector<ColNum> &maxV,
+                       const Vector<ColNum> &aggr) 
+{
+  ASSERT (toRowNum < fromRowNum);
+  StringVector& to = rows [toRowNum];
+  const StringVector& from = rows [fromRowNum];
+
+  for (const ColNum i : sum)
+  {
+    const Header& h = header [i];
+    ASSERT (h. numeric);
+    ostringstream oss;
+    ONumber on (oss, h. decimals, h. scientific);
+    oss << stod (to [i]) + stod (from [i]);
+    to [i] = oss. str ();
+  }
+
+  for (const ColNum i : minV)
+    if (compare (to, from, i) == 1)
+      to [i] = from [i];
+
+  for (const ColNum i : maxV)
+    if (compare (to, from, i) == -1)
+      to [i] = from [i];
+
+  for (const ColNum i : aggr)
+  {
+    constexpr char sep = ',';
+    if (from [i]. empty ())
+      continue;
+    if (to [i]. empty ())
+      to [i] = from [i];
+    else
+    {
+      StringVector vec (to [i], sep, true);
+      vec << from [i];
+      vec. sort ();
+      vec. uniq ();
+      to [i] = vec. toString (string (1, sep));
+    }
+  }
+}
+
+
+
+void TextTable::colNumsRow2values (const Vector<ColNum> &colNums,
+                                   RowNum row_num,
+                                   StringVector &values) const
+{
+  values. clear ();
+  values. reserve (colNums. size ());
+  const StringVector& row = rows [row_num];    
+  FFOR (ColNum, i, colNums. size ())
+    values << row [colNums [i]];
+}
+
+
+
+TextTable::RowNum TextTable::find (const Vector<ColNum> &colNums,
+                                   const StringVector &targetValues,
+                                   RowNum row_num_start) const
+{
+  ASSERT (colNums. size () == targetValues. size ());
+  ASSERT (row_num_start != no_index);
+  StringVector values;
+  FOR_START (RowNum, i, row_num_start, rows. size ())
+  {
+    colNumsRow2values (colNums, i, values);
+    if (values == targetValues)
+      return i;
+  }
+  return no_index;
+}
+
+
+
+StringVector TextTable::col2values (ColNum col) const
+{
+  QC_ASSERT (col < header. size ());
+  
+  Set<string> s;
+  for (const StringVector& row : rows)
+    if (! row [col]. empty ())
+      s << row [col];
+            
+  StringVector vec;  vec. reserve (s. size ());
+  insertAll (vec, s);
+  
+  return vec;
+}
+
+
+
+
+// TextTable::Key
+
+
+TextTable::Key::Key (const TextTable &tab,
+                     const StringVector &columns)
+: colNums (tab. columns2nums (columns))
+{
+  data. rehash (tab. rows. size ());
+  StringVector values;  
+  FFOR (RowNum, i, tab. rows. size ())
+  {
+    tab. colNumsRow2values (colNums, i, values);
+    for (const string& s : values)
+      if (s. empty ())
+        throw Error (tab, "Empty value in key, in row " + to_string (i + 1));
+    if (data. find (values) != data. end ())
+      throw Error (tab, "Duplicate key " + values. toString (",") + " for the key on " + columns. toString (","));
+    ASSERT (i != no_index);
+    data [values] = i;
+  }  
+}
+
+
+
+
+// TextTable::Index
+
+
+TextTable::Index::Index (const TextTable &tab,
+                         const StringVector &columns)
+: colNums (tab. columns2nums (columns))
+{
+  data. rehash (tab. rows. size ());
+  StringVector values;  
+  FFOR (RowNum, i, tab. rows. size ())
+  {
+    tab. colNumsRow2values (colNums, i, values);
+    data [values] << i;
+  }  
+}
+
+
+
+
+}
+
+
